+++ bluefin-internal/CHANGELOG.md
@@ -1,3 +1,7 @@
+* Make `Vault`'s `Key` type role representational.  This is
+  technically a PVP violation but since it's fixing a type safety bug
+  we're not going to release a major version for it.
+
 # 0.9.1.0
 
 * Add `yieldToPureList`